Removed the strips for when you try to login, put is_valid_email in its own file and made it fully whitelisting
